refactor: rename stampCollecting to trailMap
